Create HTML & CSS Study III
It includes the code for making a awesome Landing Page of a Responsive Website. It is responsive website in nature means that it will change its layout according to the size of the display screen i.e. different for mobile screens and different for computers. It includes only HTML (Hyper Text Markup Language) and CSS (Cascading Style Sheets).
It includes HTML & CSS Study III code of Progate HTML and CSS course.