Create Array In Objects
Complete the part of the for loop between the () by following the instructions below:
[Variables]
• Define a variable i and assign the value 0 to it
[Condition]
• Using the length method, add a condition that is true when "the value of the i variable is less than the number of elements in characters"
[Variable Update]
• Add 1 to the i variable
Return to Slides
Next let's write code for the inside of the for loop.
Define the 
character
 constant and assign it the element at index i of the characters array.
1)
An array is assigned to the characters constant. The for loop assigns each element of that array to character one by one in order.
Be careful about whether or not there should be an s at the end of the constant names. 
2)
"The element of the characters array with the index i" can be written as follows:
characters[i]
Assign this value to the character constant.
＊You can use “const” inside the for loop because each loop has a different scope.
Print the output 
My name is ____
. But, the blank (____) should be replaced with the value of the character constant's name property.
Either combining strings (concatenation) or template literals can be used to print the output here.
script.js
Print the output 
I am ____ years old
. But, the blank (____) should be replaced by the value of the character constant's age property.