Create Class Inheritance Overridding
Inheritance is a way of creating a new class based on the existing class. When we create a class using inheritance we use extends.
When a method that has the same name as the method in the parent class is defined in the child class the child class method will be used . This property is referred to as overridding in JavaScript.