Python Study III
Let's let the user pick a hand using input().
We'll be using numbers to get the input, and print the corresponding hand.
Below the line where the player_name variable is defined, print 
Pick a hand: (0: Rock, 1: Paper, 2: Scissors)
.
Now let's get the input. ・Using input(), get a value from the user after printing 
Please enter a number (0-2): 
.
・Assign the value to the  player_hand
 variable after converting it to an integer using int().

Let's use the input to call the print_hand function.
script.py
Change the first argument for the print_hand function to 
player_hand
.

We now have to rewrite the print_hand function to print hands  based on the integer value.

Inside the print_hand function, assign the following list of hands to the hands variable:
Rock

Paper

Scissors

The hand parameter has the integer value received from the user's input.
Get the element from the list hands that has the index number equal to hand, and use it to update the print() function on line 6.